requirements.txt
streamlit
pandas
plotly
numpy
python-dateutil
pytz
psycopg2-binary
